Functional Build and Test scripts for Ubuntu Debian package
TODO: Release script, push this repository public, and push package upstream!

+++ ubuntu/build.sh
@@ -1,31 +1,53 @@
 #!/bin/bash
 
-# Here is my attempt to automate building a .deb package from the source
+# Building a .deb/.dsc package from the release-source
 
-### Configs ###
-source configs.sh
+
+source config.sh
+
 
 ### Preliminary setup ###
 
+# Argument parsing
+configParse "$@"
+
+# Setup variables
+configSetup
+
+
+cd ${UBUNTU_RELEASE}
+
+
+### Build the tarballs ###
+
 # Create gzipped tarball from source
 rm -f ${TARBALL}
-tar -cvzf ${TARBALL} ${SRC}
+tar -cvzf ${TARBALL} -C $(dirname ${SRC}) $(basename ${SRC})
 
 # Generate template via bzr
 rm -f ${ORIG_TARBALL}
-( cd ${PKG} && bzr remove-branch --force . )
-rm -rf ${PKG}
+
+# Double check 'rm -rf ${VAR}' (could cost you your life (or home directory))
+if [ -z ${PKG} ]; then
+    echo "Critical Error: \${PKG} is not defined"
+    exit 1
+fi
+
+if [ -d "${PKG}" ]; then
+    ( cd ${PKG} && bzr remove-branch --force . )
+    rm -rf ${PKG}
+fi
+
 bzr dh-make ${PKG} ${VERSION} ${TARBALL} <<EOD
 s
 
 EOD
 
-### debian files ###
+
+### ${PKG}/debian files ###
 # Overwrite generated debian directory with custom debian files
 rm -rf ${PKG}/debian
-#( cd ${PKG} &&  bzr revert )
 cp -r ${DEB_SRC} ${PKG}/debian
-rm -rf ${PKG}/debian/tmp # Temporary
 
 # Pre-clean
 rm -f ${CHANGES}
@@ -36,11 +58,9 @@ rm -rf build-area
 
 
 ### Bazaar & Build ###
+
 # Generate debian files to repository
 ( cd ${PKG} && bzr add debian && bzr commit -m "Initial commit of Debian Packaging" )
 
 # Build package!
 ( cd ${PKG} && bzr builddeb -- -us -uc )
-
-### Information about package ###
-#lesspipe ${DEB}
